POPUG-HW-02 Review: refactored CUD Task events into business events.
