+++ task-tracker/src/main/java/org/uber/popug/task/tracker/config/TaskCompletionServicesConfig.java
@@ -2,7 +2,7 @@ package org.uber.popug.task.tracker.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.uber.popug.task.tracker.kafka.producer.TasksCUDEventProducer;
+import org.uber.popug.task.tracker.kafka.producer.TasksBusinessEventProducer;
 import org.uber.popug.task.tracker.mapping.TasksKafkaEventMapper;
 import org.uber.popug.task.tracker.repository.TaskRepository;
 import org.uber.popug.task.tracker.repository.UserRepository;
@@ -28,14 +28,14 @@ public class TaskCompletionServicesConfig {
             TaskRepository taskRepository,
             UserRepository userRepository,
             TasksKafkaEventMapper tasksKafkaEventMapper,
-            TasksCUDEventProducer tasksCUDEventProducer
+            TasksBusinessEventProducer tasksBusinessEventProducer
     ) {
         return new TaskCompletionServiceImpl(
                 taskMembershipCheckingService,
                 taskRepository,
                 userRepository,
                 tasksKafkaEventMapper,
-                tasksCUDEventProducer
+                tasksBusinessEventProducer
         );
     }
 }
