+++ task-tracker/src/main/java/org/uber/popug/task/tracker/kafka/producer/event/business/TaskCompletedEvent.java
@@ -1,4 +1,4 @@
-package org.uber.popug.task.tracker.kafka.producer.dto;
+package org.uber.popug.task.tracker.kafka.producer.event.business;
 
 import jakarta.annotation.Nonnull;
 import org.apache.kafka.clients.producer.ProducerRecord;
@@ -10,33 +10,34 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
-public record TaskCreatedEvent(
+public record TaskCompletedEvent(
         @Nonnull UUID publicId,
-        @Nonnull UUID publicAssigneeId,
+        @Nonnull UUID extPublicAssigneeId,
         @Nonnull String description,
-        @Nonnull LocalDateTime creationDate
+        @Nonnull LocalDateTime completionDate
 ) implements KafkaProducerRecordEvent<String, Object> {
-    private static final List<RecordHeader> TASK_CREATED_EVENT_TYPE = List.of(
-            new RecordHeader("type", "task.created".getBytes(StandardCharsets.UTF_8)),
+
+    private static final List<RecordHeader> TASK_COMPLETED_EVENT_TYPE = List.of(
+            new RecordHeader("type", "task.completed".getBytes(StandardCharsets.UTF_8)),
             new RecordHeader("version", "1.0".getBytes(StandardCharsets.UTF_8)),
             new RecordHeader("producer", "uber-popug.task-tracker".getBytes(StandardCharsets.UTF_8))
     );
 
     @Override
     public ProducerRecord<String, Object> asProducerRecord(String topicName) {
-        final var taskCreatedEventProducerRecord = new ProducerRecord<>(
+        final var taskCompletedEventProducerRecord = new ProducerRecord<>(
                 topicName,
                 recordKey(),
                 recordValue()
         );
-        TASK_CREATED_EVENT_TYPE.forEach(recordHeader -> taskCreatedEventProducerRecord.headers().add(recordHeader));
+        TASK_COMPLETED_EVENT_TYPE.forEach(recordHeader -> taskCompletedEventProducerRecord.headers().add(recordHeader));
 
-        return taskCreatedEventProducerRecord;
+        return taskCompletedEventProducerRecord;
     }
 
     @Override
     public String recordKey() {
-        return publicId().toString();
+        return publicId.toString();
     }
 
     @Override
