+++ task-tracker/src/main/java/org/uber/popug/task/tracker/service/impl/TaskAddingServiceImpl.java
@@ -1,7 +1,7 @@
 package org.uber.popug.task.tracker.service.impl;
 
 import lombok.RequiredArgsConstructor;
-import org.uber.popug.task.tracker.kafka.producer.TasksCUDEventProducer;
+import org.uber.popug.task.tracker.kafka.producer.TasksBusinessEventProducer;
 import org.uber.popug.task.tracker.mapping.TasksDtoMapper;
 import org.uber.popug.task.tracker.repository.TaskRepository;
 import org.uber.popug.task.tracker.rest.generated.model.PostAddTaskRequestDto;
@@ -15,14 +15,14 @@ public class TaskAddingServiceImpl implements TaskAddingService {
     private final TasksDtoMapper tasksDtoMapper;
     private final TaskAssignmentService taskAssignmentService;
     private final TaskRepository taskRepository;
-    private final TasksCUDEventProducer tasksCUDEventProducer;
+    private final TasksBusinessEventProducer tasksBusinessEventProducer;
 
     @Override
     public PostAddTaskResponseDto addNewTask(PostAddTaskRequestDto postTasksRequestDto) {
         final var taskForCreation = tasksDtoMapper.postTaskRequestsDtoToBusiness(postTasksRequestDto);
         final var assignedTask = taskAssignmentService.assignNewTask(taskForCreation);
         taskRepository.add(assignedTask);
-        tasksCUDEventProducer.sendTaskCreationEvent(assignedTask);
+        tasksBusinessEventProducer.sendTaskCreationEvent(assignedTask);
 
         return tasksDtoMapper.postTasksResponseDtoFromBusiness(assignedTask);
     }
